🛠️fix: Achievement 조회 로직 버그수정 (#51)
- type 안맞던것들 수정

+++ src/main/java/kr/modernworld/modernworldv2/growth/application/achievement/dto/AchievementDTO.java
@@ -5,7 +5,7 @@ public record AchievementDTO(
     String title,
     String description,
     String level,
-    Integer point
+    Long point
 ) {
 
 }
